Move SBML tests back to tests/sbml
