+++ .github/workflows/test_sbml_semantic_test_suite_jax.yml
@@ -10,8 +10,8 @@ on:
       - .github/workflows/test_sbml_semantic_test_suite_jax.yml
       - python/sdist/amici/jax/**
       - scripts/run-SBMLTestsuiteJax.sh
-      - tests/sbml/tests/testSBMLSuiteJax.py
-      - tests/sbml/tests/conftest.py
+      - tests/sbml/testSBMLSuiteJax.py
+      - tests/sbml/conftest.py
   check_suite:
    types: [requested]
   workflow_dispatch:
