+++ scripts/run-SBMLTestsuite.sh
@@ -29,5 +29,5 @@ if [[ -d "${RESULT_DIR}" ]]; then
 fi
 mkdir "${RESULT_DIR}"
 
-pytest ./tests/sbml/tests/testSBMLSuite.py $cases -rfsE -n auto \
+pytest ./tests/sbml/testSBMLSuite.py $cases -rfsE -n auto \
   --cov=amici --cov-report=xml:"coverage_SBMLSuite.xml" --cov-append
